backend/requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
pytz==2021.1
sqlparse==0.4.1
django-cors-headers==3.7.0
django-environ==0.4.5